Add files via upload
This interactive program asks the user about a car's information, and then prints out the details of the car on the screen. Afterwards, the program asks the users again for input. The user enters a trip information they would like to take with the car, and then at the end prints out a graph with the information of the trip.